==> notebooks/01_scrape_and_save_data.ipynb <==
# Scrape and Save Data
### Description:
This file runs the UnderstatScraper found in `src` to retrieve the English Premier League results for all 'home' matches and all 'away' matches and saves them top raw data csvs in the `data/raw/` directory.

All probability calculations depend on this data being up to date.

### Instructions: 
Only one cell! 'Run all' or 'shift + enter' to scrape the current Premier League table

Confirm that `../data/raw/away_table.csv` and `../data/raw/away_table.csv` are updated.

When finished, procede to `02_process_data.ipynb`


import sys
sys.path.append('../src')

from understat_api.understat_scraper import UnderstatScraperOrchestrator

scraper = UnderstatScraperOrchestrator()

scraper.scrape_all_to_csv()
