==> notebooks/03a_retrieve_odds.ipynb.ipynb <==
import sys
sys.path.append('../src')

from config import ODDS_API_KEY, BASE_URL
from odds_api.odds_fetcher import OddsFetcher
from odds_api.utils import parse_odds_data
import json

fetcher = OddsFetcher(api_key=ODDS_API_KEY, base_url=BASE_URL)
data = fetcher.fetch_odds(matchday="2024-12-21")
# print(json.dumps(data, indent=2))

print(fetcher.api_usage)

parse_odds_data(data)
